demo-tvOS Separated the cell and model to different files, also fixed typo on the MovieDetail folder
